applying changes
