+++ autograder/steps/pre_flight_step.py
@@ -1,9 +1,11 @@
 import logging
+from typing import List
 
 from autograder.models.abstract.step import Step
 from autograder.models.pipeline_execution import PipelineExecution
 from autograder.models.dataclass.step_result import StepResult, StepName
 from autograder.services.pre_flight_service import PreFlightService
+from autograder.services.sandbox_service import SandboxService
 
 logger = logging.getLogger(__name__)
 
@@ -12,11 +14,11 @@ class PreFlightStep(Step):
     """
     The Pre-flight step is responsible for:
         - Running Pre-Grading validations on submissions
-        - Sandboxing Submission Code (If the grading process requires executing submission code)
+        - Executing setup commands (e.g., compilation) in the sandbox
 
     Pre-Grading Checks are run in order:
     1. Required files check
-    2. Setup commands check (only if files check passes)
+    2. Setup commands check (only if files check passes and sandbox exists)
 
     If any check fails, the step returns a FAIL status with error details.
     """
@@ -24,7 +26,7 @@ class PreFlightStep(Step):
     def __init__(self, setup_config):
         self._setup_config = setup_config
         self._pre_flight_service = None
-        # Don't create service here, create it per-execution with language
+        self._sandbox_service = SandboxService()
 
     @property
     def step_name(self) -> StepName:
@@ -32,23 +34,20 @@ class PreFlightStep(Step):
 
     def _execute(self, pipeline_exec: PipelineExecution) -> PipelineExecution:
         """
-        Execute pre-flight checks (required files) on the submission.
-
-        Args:
-            pipeline_exec: PipelineExecution context.
-
-        Returns:
-            PipelineExecution with updated step results.
+        Execute pre-flight checks (required files) and setup commands.
+        
+        Note: Sandbox must have been created in a previous step (SandboxStep).
         """
         submission_language = pipeline_exec.submission.language
         self._pre_flight_service = PreFlightService(self._setup_config, submission_language)
 
         logger.info(
-            "Pre-flight (file check) started: external_user_id=%s, language=%s",
+            "Pre-flight checks started: external_user_id=%s, language=%s",
             pipeline_exec.submission.user_id,
             submission_language.value if submission_language else "none",
         )
 
+        # 1. Check required files
         if self._pre_flight_service.required_files:
             logger.info("Checking required files for submission (external_user_id=%s)", pipeline_exec.submission.user_id)
             files_ok = self._pre_flight_service.check_required_files(pipeline_exec.submission.submission_files)
@@ -62,12 +61,41 @@ class PreFlightStep(Step):
                     error_data=self._pre_flight_service.fatal_errors
                 ))
 
-        logger.info("Pre-flight (file check) passed (external_user_id=%s)", pipeline_exec.submission.user_id)
+        # 2. Check setup commands (requires sandbox from a previous step)
+        if self._pre_flight_service.setup_commands:
+            sandbox = pipeline_exec.sandbox
+            if not sandbox:
+                # This check ensures that if setup commands exist, we must have an active sandbox.
+                # If SandboxStep was skipped but we have commands, we must report an error.
+                logger.error("Sandbox required for setup commands but was not found in pipeline execution.")
+                return pipeline_exec.add_step_result(StepResult.fail(
+                    step=self.step_name,
+                    error="Sandbox environment is missing for setup command execution."
+                ))
+
+            logger.info("Running setup commands in sandbox (external_user_id=%s)", pipeline_exec.submission.user_id)
+            setup_ok = self._sandbox_service.run_setup_commands(sandbox, self._pre_flight_service.setup_commands)
+            
+            if not setup_ok:
+                error_msg = self._format_errors()
+                logger.warning("Setup commands failed (external_user_id=%s): %s", pipeline_exec.submission.user_id, error_msg)
+                return pipeline_exec.add_step_result(StepResult.fail(
+                    step=self.step_name,
+                    error=error_msg,
+                    error_data=self._sandbox_service.fatal_errors
+                ))
+
+        logger.info("Pre-flight checks passed (external_user_id=%s)", pipeline_exec.submission.user_id)
         return pipeline_exec.add_step_result(StepResult.success(self.step_name, None))
 
     def _format_errors(self) -> str:
-        """Format all preflight errors into a single error message."""
-        if not self._pre_flight_service.has_errors():
+        """Format all errors from both services into a single message."""
+        errors = []
+        if self._pre_flight_service and self._pre_flight_service.has_errors():
+            errors.extend(self._pre_flight_service.get_error_messages())
+        if self._sandbox_service and self._sandbox_service.has_errors():
+            errors.extend(self._sandbox_service.get_error_messages())
+            
+        if not errors:
             return "Unknown preflight error"
-        error_messages = self._pre_flight_service.get_error_messages()
-        return "\n".join(error_messages)
+        return "\n".join(errors)
